refactor: break down cache fast_path.rs and warming.rs into focused modules
- Split fast_path.rs (394 lines) into 6 modules (largest: 150 lines)
- Split warming.rs (378 lines) into 7 modules (largest: 104 lines)
- Use directory hierarchies with hyphens, no underscores
- Group by function: core ops, specialized types, patterns, candidates
- Maintain all functionality and tests
